Move `UserRole` to `users.schemas`.
