+++ app/components/users/schemas.py
@@ -1,5 +1,5 @@
 import datetime
-from typing import Optional
+from typing import Optional, List
 
 from pydantic import BaseModel, EmailStr, validator
 
@@ -59,3 +59,13 @@ class UserOut(UserRepresentation):
     role: str
     # TODO REMOVE
     registration_token: Optional[str] = None
+
+
+# TODO: Unused as of 31/01/2023
+class UserRole(BaseModel):
+
+    verbose_name: str
+    permissions: List[str]
+
+    class Config:
+        orm_mode = True
